refactor: 팩토리 클래스 삭제
